requirements.txt
customtkinter==5.2.2
darkdetect==0.8.0
packaging==24.1
psycopg2-binary==2.9.9
python-dotenv==1.0.1
